More info on moving back to stat command
Reverted back to using stat command for logged in user due to mismatches in values returned when checking admin group members against logged in user.
e.g. System Config had the username with an uppercase first letter for the surname but the username listed in the admin group was all lowercase.
This caused false negatives when the post check was completed. admin rights had been provided but it returned an error incorrectly.